+++ src/device_node.h
@@ -30,13 +30,13 @@ class DeviceNode {
 
   void add_subscriber();
   void remove_subscriber();
-  uint8_t subscriber_count() const { return subscriber_count_; }
+  uint16_t subscriber_count() const { return subscriber_count_; }
   bool has_anonymous_subscriber() const { return has_anonymous_subscriber_; }
   void mark_anonymous_subscriber() { has_anonymous_subscriber_ = true; }
 
   void add_publisher();
   void remove_publisher();
-  uint8_t publisher_count() const { return publisher_count_; }
+  uint16_t publisher_count() const { return publisher_count_; }
   bool has_anonymous_publisher() const { return has_anonymous_publisher_; }
   void mark_anonymous_publisher() { has_anonymous_publisher_ = true; }
 
@@ -99,9 +99,9 @@ class DeviceNode {
 
   mutable base::Mutex lock_{};
 
-  uint8_t subscriber_count_{0};
+  uint16_t subscriber_count_{0};
   bool has_anonymous_subscriber_{false};
-  uint8_t publisher_count_{0};
+  uint16_t publisher_count_{0};
   bool has_anonymous_publisher_{false};
 
   intrusive_list::forward_list<detail::ReceiverBase, &detail::ReceiverBase::receiver_node> receiver_list_;
